Run tests in a Docker container on Travis
The main advantage of this approach is that there is more control over
the environment, and is much easier to replicate locally when things go
wrong.

Sidebar: I've had very little luck getting Debian Jessie to work due to
errors running systemd in a container. By all accounts, I should be able
to get this working (Jeff Geerling seems to have it working on CentOS).