Add custom init file for Ubuntu
The version provided by the distribution behaves incorrectly if we
attempt to start again. This ensures that icecast creates a pid file on
start, and that the `start-stop-daemon` tool checks this file.

This change should probably be pushed upstream, but Ubuntu is currently
only a testing target for me (on Travis) due to limitations I'm
experiencing with running systemd in a docker container.